requirements.txt
flask
flask-cors
PyPDF2
docx2txt
python-docx
PyMuPDF